+++ HISTORY.rst
@@ -21,6 +21,7 @@ Latest (Unreleased)
 
 **Features and Improvements**
 * BSIBSO-1121: Active field in 'purchase.order', filter into view
+* BSIBSO-1117: have a customer dependent numbering for analytic accounts
 
 **Bugfixes**