+++ HISTORY.rst
@@ -25,6 +25,7 @@ Latest (Unreleased)
   to have only purchase orders in the currency of the invoice
 * BSIBSO-1121: Active field in 'purchase.order', filter into view
 * BSIBSO-1117: have a customer dependent numbering for analytic accounts
+* bso_backbone: change the visibility of some fields in the views
 
 **Bugfixes**